# Meizon-Inventory-Swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Inventory Management Service",
    "version": "1.0.1",
    "title": "Meizon Inventory Service",
    "license": {
      "name": "Apache Licence Version 2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/api/v1/location/{locationId}": {
      "get": {
        "tags": [
          "location-controller-v-1"
        ],
        "summary": "getLocationByLocationId",
        "operationId": "getLocationByLocationIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "description": "locationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "location-controller-v-1"
        ],
        "summary": "editLocation",
        "operationId": "editLocationUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "locationDto",
            "description": "locationDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LocationDto"
            }
          },
          {
            "name": "locationId",
            "in": "path",
            "description": "locationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "location-controller-v-1"
        ],
        "summary": "deleteByLocationId",
        "operationId": "deleteByLocationIdUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "description": "locationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/payment/": {
      "get": {
        "tags": [
          "payment-controller-v-1"
        ],
        "summary": "getAllPayments",
        "operationId": "getAllPaymentsUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "payment-controller-v-1"
        ],
        "summary": "makePayment",
        "operationId": "makePaymentUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "payments",
            "description": "payments",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PaymentDto"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/payment/customer-info": {
      "get": {
        "tags": [
          "payment-controller-v-1"
        ],
        "summary": "getPaymentByCustomeInfo",
        "operationId": "getPaymentByCustomeInfoUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "companyName",
            "in": "query",
            "description": "companyName",
            "required": true,
            "type": "string"
          },
          {
            "name": "firstName",
            "in": "query",
            "description": "firstName",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastName",
            "in": "query",
            "description": "lastName",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/payment/{paymentId}": {
      "get": {
        "tags": [
          "payment-controller-v-1"
        ],
        "summary": "getPaymentById",
        "operationId": "getPaymentByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "description": "paymentId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "payment-controller-v-1"
        ],
        "summary": "deleteByLocationId",
        "operationId": "deleteByLocationIdUsingDELETE_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "description": "paymentId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "AddressDto": {
      "type": "object",
      "properties": {
        "street": {
          "type": "string"
        },
        "unitNumber": {
          "type": "string"
        }
      },
      "title": "AddressDto"
    },
    "ContactDto": {
      "type": "object",
      "properties": {
        "detail": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "title": "ContactDto"
    },
    "CurrencyDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      },
      "title": "CurrencyDto"
    },
    "CustomerDto": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/AddressDto"
        },
        "city": {
          "type": "string"
        },
        "companyName": {
          "type": "string"
        },
        "contacts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContactDto"
          }
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "region": {
          "type": "string"
        }
      },
      "title": "CustomerDto"
    },
...
    "LocationDto": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/AddressDto"
        },
        "description": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "province": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "title": "LocationDto"
    },
    "PaymentDto": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "number",
          "format": "double"
        },
        "currency": {
          "$ref": "#/definitions/CurrencyDto"
        },
        "customer": {
          "$ref": "#/definitions/CustomerDto"
        },
        "paymentHistoryId": {
          "type": "integer",
          "format": "int64"
        },
        "paymentTime": {
          "type": "string",
          "format": "date-time"
        },
        "paymentType": {
          "$ref": "#/definitions/PaymentTypeDto"
        }
      },
      "title": "PaymentDto"
    },
    "PaymentTypeDto": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "typeId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "PaymentTypeDto"
    },
...
  }
}